Renable block configuration deletion

+++ react/components/EditorContainer/Sidebar/BlockEditor/BlockConfigurationList/hooks.ts
@@ -1,9 +1,30 @@
 import { useCallback } from 'react'
+import { defineMessages } from 'react-intl'
 
 import { useEditorContext } from '../../../../EditorContext'
 import { getIsDefaultContent } from '../utils'
 
 import { UseListHandlers } from './typings'
+import { getDeleteStoreUpdater } from './utils'
+
+const messages = defineMessages({
+  deleteError: {
+    defaultMessage: 'Something went wrong. Please try again.',
+    id: 'admin/pages.editor.components.content.delete.error',
+  },
+  deleteSuccess: {
+    defaultMessage: 'Content deleted.',
+    id: 'admin/pages.editor.components.content.delete.success',
+  },
+  resetError: {
+    defaultMessage: 'Something went wrong. Please try again.',
+    id: 'admin/pages.editor.components.content.reset.error',
+  },
+  resetSuccess: {
+    defaultMessage: 'Content reset.',
+    id: 'admin/pages.editor.components.content.reset.success',
+  },
+})
 
 export const useListHandlers: UseListHandlers = ({
   deleteContent,
@@ -34,48 +55,55 @@ export const useListHandlers: UseListHandlers = ({
   //       })
   //     }
 
-  // TODO
-  const handleConfigurationDeletion = useCallback(
+  const handleConfigurationDelete = useCallback(
     async (configuration: ExtensionConfiguration) => {
       editor.setIsLoading(true)
 
       const action = getIsDefaultContent(configuration) ? 'reset' : 'delete'
 
-      try {
-        await deleteContent({
-          variables: {
-            contentId: configuration.contentId,
-            pageContext: iframeRuntime.route.pageContext,
-            template: editor.blockData.template,
-            treePath: editor.blockData.serverTreePath,
-          },
-        })
+      const {
+        blockData: { id: blockId, serverTreePath, template },
+        setBlockData,
+      } = editor
 
-        // TODO
-        // await updateIframeAndSiteEditor()
+      if (blockId && serverTreePath && template) {
+        let wasSuccessful = true
 
-        editor.setIsLoading(false)
+        try {
+          await deleteContent({
+            update: getDeleteStoreUpdater({
+              blockId,
+              iframeRuntime,
+              serverTreePath,
+              setBlockData,
+              template,
+            }),
+            variables: {
+              contentId: configuration.contentId,
+              pageContext: iframeRuntime.route.pageContext,
+              template,
+              treePath: serverTreePath,
+            },
+          })
+        } catch (error) {
+          wasSuccessful = false
 
-        showToast({
-          horizontalPosition: 'right',
-          message: intl.formatMessage({
-            id: `admin/pages.editor.components.content.${action}.success`,
-          }),
-        })
-      } catch (e) {
-        editor.setIsLoading(false)
+          console.error(error)
+        } finally {
+          const messageKey = action + (wasSuccessful ? 'Success' : 'Error')
 
-        showToast({
-          horizontalPosition: 'right',
-          message: intl.formatMessage({
-            id: `admin/pages.editor.components.content.${action}.error`,
-          }),
-        })
+          editor.setIsLoading(false)
 
-        console.error(e)
+          showToast({
+            horizontalPosition: 'right',
+            message: intl.formatMessage(
+              messages[messageKey as keyof typeof messages]
+            ),
+          })
+        }
       }
     },
-    [deleteContent, editor, iframeRuntime.route.pageContext, intl, showToast]
+    [deleteContent, editor, iframeRuntime, intl, showToast]
   )
 
   const handleQuit = useCallback(
@@ -90,7 +118,7 @@ export const useListHandlers: UseListHandlers = ({
   )
 
   return {
-    handleConfigurationDeletion,
+    handleConfigurationDelete,
     handleQuit,
   }
 }
